+++ docs/conf.py
@@ -2,6 +2,7 @@
 #
 # GPIO Zero: a library for controlling the Raspberry Pi's GPIO pins
 # Copyright (c) 2015-2019 Dave Jones <dave@waveform.org.uk>
+# Copyright (c) 2019 Ben Nuttall <ben@bennuttall.com>
 # Copyright (c) 2016 Thijs Triemstra <info@collab.nl>
 #
 # Redistribution and use in source and binary forms, with or without