+++ gpiozero/exc.py
@@ -1,6 +1,7 @@
 # GPIO Zero: a library for controlling the Raspberry Pi's GPIO pins
 # Copyright (c) 2016-2019 Dave Jones <dave@waveform.org.uk>
 # Copyright (c) 2016-2019 Andrew Scheller <github@loowis.durge.org>
+# Copyright (c) 2019 Kosovan Sofiia <sofiia.kosovan@gmail.com>
 # Copyright (c) 2019 Ben Nuttall <ben@bennuttall.com>
 #
 # Redistribution and use in source and binary forms, with or without