+++ docs/changelog.rst
@@ -1,5 +1,5 @@
 .. GPIO Zero: a library for controlling the Raspberry Pi's GPIO pins
-.. Copyright (c) 2016-2019 Ben Nuttall <ben@bennuttall.com>
+.. Copyright (c) 2016-2020 Ben Nuttall <ben@bennuttall.com>
 .. Copyright (c) 2015-2019 Dave Jones <dave@waveform.org.uk>
 .. Copyright (c) 2016 Andrew Scheller <github@loowis.durge.org>
 ..